requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.5
cssselect2==0.4.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.7
django-crispy-forms==1.11.1
django-formtools==2.2
django-import-export==2.5.0
django-reversion==3.0.9
django-simpleui==2021.3
django-utils-six==2.0
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
httplib2==0.9.2
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.1
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
pypiwin32==223
python-bidi==0.4.2
pytz==2021.1
pywin32==300
PyYAML==5.4.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
WeasyPrint==52.3
webencodings==0.5.1
wincertstore==0.2
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
